linkedin_ai_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart
pdfplumber
anthropic
python-dotenv
aiofiles
psycopg2-binary
slowapi
bleach
